add basic federation manager tests
